Reuse existing files
