--- site/gatsby-site/src/components/NewRelatedIncidentsArea.js
@@ -1,134 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { Button, Spinner } from 'flowbite-react';
-import { LocalizedLink } from 'gatsby-theme-i18n';
-import { Trans, useTranslation } from 'react-i18next';
-import NewSimilaritySelector from './NewSimilaritySelector';
-import { faArrowDown } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-
-const NewRelatedIncidentsArea = ({
-  columnKey,
-  header,
-  reports,
-  incidents,
-  loading,
-  setFieldValue,
-  editSimilar = true,
-  editId = true,
-  error,
-}) => {
-  const { t } = useTranslation();
-
-  const [reportsOpened, setReportsOpened] = useState(false);
-
-  const [maxIncidents, setMaxIncidents] = useState(3);
-
-  const [similarList, setSimilarList] = useState(reports || incidents || []);
-
-  const visible = reports || incidents || loading;
-
-  const reportsExist = (reports || incidents) && !loading;
-
-  useEffect(() => {
-    if (reportsExist) {
-      setReportsOpened(true);
-    }
-  }, [reportsExist]);
-
-  useEffect(() => {
-    if (reports || incidents) {
-      setSimilarList(reports || incidents);
-    } else {
-      setSimilarList([]);
-    }
-  }, [reports, incidents]);
-
-  if (!visible) {
-    return null;
-  }
-  return (
-    <div data-cy={`related-${columnKey}`}>
-      <div key={'header'}>
-        {header}
-        {loading && <Spinner size={'sm'} />}
-      </div>
-      <div className={`${reportsOpened ? 'reports open' : 'reports'} flex flex-wrap gap-2`}>
-        {similarList &&
-          similarList.slice(0, maxIncidents).map((val) => (
-            <div
-              className="py-2 px-3 max-w-sm bg-blue-100 border-blue-200 rounded-lg border shadow-md dark:bg-gray-800 dark:border-gray-700 flex-1/3 flex flex-col justify-between"
-              key={val.url || val.incident_id}
-              data-cy="result"
-            >
-              <span>
-                {val?.incident_id && (
-                  <div className="text-lg">
-                    <LocalizedLink
-                      to={'/cite/' + val.incident_id}
-                      className="text-black hover:text-blue-700"
-                    >
-                      #{val.incident_id}
-                    </LocalizedLink>
-                  </div>
-                )}
-                <a
-                  href={val.url || '/cite/' + val.incident_id}
-                  data-cy="title"
-                  className="text-black hover:text-blue-700 text-sm"
-                >
-                  {val.title}
-                </a>
-              </span>
-              <div className="flex justify-between">
-                {setFieldValue && editSimilar && (
-                  <div>
-                    <p className="text-xs my-2">
-                      <Trans>Is this incident related?</Trans>
-                    </p>
-                    <NewSimilaritySelector incident_id={val.incident_id} />
-                  </div>
-                )}
-                {val.incident_id && setFieldValue && editId && (
-                  <div className="self-end">
-                    <Button
-                      size={'xs'}
-                      data-cy="set-id"
-                      onClick={() => setFieldValue && setFieldValue('incident_id', val.incident_id)}
-                    >
-                      <p className="m-0">
-                        <Trans>Use ID</Trans> #{val.incident_id}
-                      </p>
-                    </Button>
-                  </div>
-                )}
-              </div>
-            </div>
-          ))}
-        {!loading && (error || reports?.length == 0 || incidents?.length == 0) && (
-          <span>{error ? error : t('No related reports found.')}</span>
-        )}
-      </div>
-      {similarList.length > maxIncidents && (
-        <div className="mt-2">
-          <Button
-            size={'xs'}
-            color="light"
-            onClick={() => {
-              setMaxIncidents(maxIncidents + 3);
-            }}
-          >
-            <Trans>Load more</Trans>
-
-            <FontAwesomeIcon
-              icon={faArrowDown}
-              className="fas fa-times ml-1"
-              style={{ marginRight: '1ch' }}
-            />
-          </Button>
-        </div>
-      )}
-    </div>
-  );
-};
-
-export default NewRelatedIncidentsArea;
