+++ site/gatsby-site/src/components/RelatedIncidentsArea.js
@@ -1,61 +1,10 @@
 import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
-import { ListGroup, Button, ButtonToolbar } from 'react-bootstrap';
-import { Spinner } from 'flowbite-react';
-import SimilaritySelector from './SimilaritySelector';
+import { Button, Spinner } from 'flowbite-react';
 import { LocalizedLink } from 'gatsby-theme-i18n';
 import { Trans, useTranslation } from 'react-i18next';
-
-const ListContainer = styled(ListGroup)`
-  max-height: 0px;
-  overflow: hidden;
-  transition: max-height 1s ease-in;
-
-  .reports {
-    max-height: 0px;
-    &.open {
-      max-height: 33.3333vh;
-      overflow-y: auto;
-    }
-    overflow: hidden;
-    transition: max-height 0.5s ease-in;
-  }
-
-  &.open {
-    margin: 1em 0;
-    max-height: 100vh;
-    overflow-y: auto;
-  }
-`;
-
-const ReportRow = styled(ListGroup.Item)`
-  display: flex !important;
-  align-items: center;
-  > *:first-child {
-    flex-shrink: 1;
-    margin-right: auto;
-  }
-  button.set-id {
-    margin-left: 1ch;
-    flex-shrink: 0 !important;
-    width: 8em;
-  }
-  @media (max-width: 860px) {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 0.5em;
-  }
-`;
-
-const ReportToolbar = styled(ButtonToolbar)`
-  flex-shrink: 0;
-  align-items: center;
-`;
-
-const SelectorLabel = styled.label`
-  margin-left: 1ch;
-  margin-right: 1ch;
-`;
+import SimilaritySelector from './SimilaritySelector';
+import { faArrowDown } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 const RelatedIncidentsArea = ({
   columnKey,
@@ -70,19 +19,15 @@ const RelatedIncidentsArea = ({
 }) => {
   const { t } = useTranslation();
 
-  const [listOpened, setListOpened] = useState(false);
-
   const [reportsOpened, setReportsOpened] = useState(false);
+
+  const [maxIncidents, setMaxIncidents] = useState(3);
+
+  const [similarList, setSimilarList] = useState(reports || incidents || []);
 
   const visible = reports || incidents || loading;
 
   const reportsExist = (reports || incidents) && !loading;
-
-  useEffect(() => {
-    if (visible) {
-      setListOpened(true);
-    }
-  }, [visible]);
 
   useEffect(() => {
     if (reportsExist) {
@@ -90,60 +35,98 @@ const RelatedIncidentsArea = ({
     }
   }, [reportsExist]);
 
+  useEffect(() => {
+    if (reports || incidents) {
+      setSimilarList(reports || incidents);
+    } else {
+      setSimilarList([]);
+    }
+  }, [reports, incidents]);
+
   if (!visible) {
     return null;
   }
   return (
-    <div className="bootstrap">
-      <ListContainer data-cy={`related-${columnKey}`} className={listOpened ? 'open' : ''}>
-        <ListGroup.Item variant="secondary" key={'header'} className="flex gap-1">
-          {header}
-          {loading && <Spinner size={'sm'} />}
-        </ListGroup.Item>
-        <div className={reportsOpened ? 'reports open' : 'reports'}>
-          {(reports || incidents) &&
-            (reports || incidents).map((val) => (
-              <ReportRow key={val.url || val.incident_id} data-cy="result">
-                <span>
-                  {val?.incident_id && (
-                    <>
-                      <LocalizedLink to={'/cite/' + val.incident_id}>
-                        #{val.incident_id}
-                      </LocalizedLink>{' '}
-                      –{' '}
-                    </>
-                  )}
-                  <a href={val.url || '/cite/' + val.incident_id} data-cy="title">
-                    {val.title}
-                  </a>
-                </span>
-                <ReportToolbar>
-                  {setFieldValue && editSimilar && (
-                    <>
-                      <SelectorLabel htmlFor="similar-selector">
-                        <Trans>Related:</Trans>
-                      </SelectorLabel>
-                      <SimilaritySelector incident_id={val.incident_id} />
-                    </>
-                  )}
-                  {val.incident_id && setFieldValue && editId && (
-                    <Button
-                      data-cy="set-id"
-                      className="set-id"
-                      onClick={() => setFieldValue && setFieldValue('incident_id', val.incident_id)}
-                      style={{ whiteSpace: 'nowrap' }}
+    <div data-cy={`related-${columnKey}`}>
+      <div key={'header'}>
+        {header}
+        {loading && <Spinner size={'sm'} />}
+      </div>
+      <div className={`${reportsOpened ? 'reports open' : 'reports'} flex flex-wrap gap-2`}>
+        {similarList &&
+          similarList.slice(0, maxIncidents).map((val) => (
+            <div
+              className="py-2 px-3 max-w-sm bg-blue-100 border-blue-200 rounded-lg border shadow-md dark:bg-gray-800 dark:border-gray-700 flex-1/3 flex flex-col justify-between"
+              key={val.url || val.incident_id}
+              data-cy="result"
+            >
+              <span>
+                {val?.incident_id && (
+                  <div className="text-lg">
+                    <LocalizedLink
+                      to={'/cite/' + val.incident_id}
+                      className="text-black hover:text-blue-700"
                     >
-                      <Trans>Use ID</Trans> <span className="incident-id">#{val.incident_id}</span>
+                      #{val.incident_id}
+                    </LocalizedLink>
+                  </div>
+                )}
+                <a
+                  href={val.url || '/cite/' + val.incident_id}
+                  data-cy="title"
+                  className="text-black hover:text-blue-700 text-sm"
+                >
+                  {val.title}
+                </a>
+              </span>
+              <div className="flex justify-between">
+                {setFieldValue && editSimilar && (
+                  <div>
+                    <p className="text-xs my-2">
+                      <Trans>Is this incident related?</Trans>
+                    </p>
+                    <SimilaritySelector incident_id={val.incident_id} />
+                  </div>
+                )}
+                {val.incident_id && setFieldValue && editId && (
+                  <div className="self-end">
+                    <Button
+                      size={'xs'}
+                      data-cy="set-id"
+                      onClick={() => setFieldValue && setFieldValue('incident_id', val.incident_id)}
+                    >
+                      <p className="m-0">
+                        <Trans>Use ID</Trans> #{val.incident_id}
+                      </p>
                     </Button>
-                  )}
-                </ReportToolbar>
-              </ReportRow>
-            ))}
-          {!loading && (error || reports?.length == 0 || incidents?.length == 0) && (
-            <ListGroup.Item>{error ? error : t('No related reports found.')}</ListGroup.Item>
-          )}
+                  </div>
+                )}
+              </div>
+            </div>
+          ))}
+        {!loading && (error || reports?.length == 0 || incidents?.length == 0) && (
+          <span>{error ? error : t('No related reports found.')}</span>
+        )}
+      </div>
+      {similarList.length > maxIncidents && (
+        <div className="mt-2">
+          <Button
+            size={'xs'}
+            color="light"
+            onClick={() => {
+              setMaxIncidents(maxIncidents + 3);
+            }}
+          >
+            <Trans>Load more</Trans>
+
+            <FontAwesomeIcon
+              icon={faArrowDown}
+              className="fas fa-times ml-1"
+              style={{ marginRight: '1ch' }}
+            />
+          </Button>
         </div>
-      </ListContainer>
+      )}
     </div>
   );
 };
